requirements.txt
Flask==2.3.3
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7 